trying

+++ streams/block/ciphers/des/triple_des.cpp
@@ -5,7 +5,7 @@
 #include <stdexcept>
 
 namespace block {
-void triple_des::keysetup(const std::uint8_t* key, const std::uint32_t keysize) {
+void triple_des::keysetup(const std::uint8_t* key, const std::uint64_t keysize) {
     std::copy_n(key, keysize, _ctx.key);
     if (_ctx.en) {
         three_des_key_setup(_ctx.key, _ctx.schedule, DES_ENCRYPT);
@@ -14,17 +14,15 @@ void triple_des::keysetup(const std::uint8_t* key, const std::uint32_t keysize) 
     }
 }
 
-void triple_des::ivsetup(const std::uint8_t* iv, const std::uint32_t ivsize) {
+void triple_des::ivsetup(const std::uint8_t* iv, const std::uint64_t ivsize) {
     throw std::runtime_error("not implemented yet");
 }
 
-void triple_des::encrypt(const std::uint8_t* plaintext,
-             std::uint8_t* ciphertext) {
-    three_des_crypt(plaintext, ciphertext, _ctx.schedule, _rounds);
+void triple_des::encrypt(const std::uint8_t* plaintext, std::uint8_t* ciphertext) {
+    three_des_crypt(plaintext, ciphertext, _ctx.schedule, unsigned(_rounds));
 }
 
-void triple_des::decrypt(const std::uint8_t* ciphertext,
-             std::uint8_t* plaintext) {
-    three_des_crypt(ciphertext, plaintext, _ctx.schedule, _rounds);
+void triple_des::decrypt(const std::uint8_t* ciphertext, std::uint8_t* plaintext) {
+    three_des_crypt(ciphertext, plaintext, _ctx.schedule, unsigned(_rounds));
 }
 } //namespace block
